:memo: update documentation to use `docker compose`
Updates the documentation to reflect the usage of `docker compose`
instead of `docker-compose`.
This change ensures that the documentation is up-to-date with the
current standard tool. It also addresses an issue where file permissions
were not being correctly set when executing commands within the `uwsgi`
container, by explicitly defining the user and group to run the command.